moved service tasks to buildSets
